Create integrations_test/ and move that hit DB there
