+++ integration_test/pg/all_test.exs
@@ -1,0 +1,1 @@
+Code.require_file "../cases/flop_test.exs", __DIR__
