--- test/support/repo.ex
@@ -1,5 +1,0 @@
-defmodule Flop.Repo do
-  use Ecto.Repo,
-    otp_app: :flop,
-    adapter: Ecto.Adapters.Postgres
-end
